moving getMatchingRecords into a hook
