+++ constants/filter.ts
@@ -45,12 +45,3 @@ export const categoryCopy: CopyHolder = {
     viewResults: 'Ver resultados',
   },
 }
-export const getMatchingRecords = (allRecords, keywords) => {
-  const filteredRecords = allRecords.records.filter(record =>
-    keywords.some(keyword => {
-      let orgTags = record.fields.org_tags.map(item => item.toLowerCase())
-      return orgTags.includes(keyword.toLowerCase())
-    }),
-  )
-  return { records: filteredRecords, category: allRecords.category }
-}
